# use pause asm insn in busyloop to run the CPU (13600K) 10 °C cooler