--- a/ggml.c
+++ b/ggml.c
@@ -11656,11 +11656,8 @@ typedef int ggml_lock_t;
 
 #define ggml_lock_init(x)    UNUSED(x)
 #define ggml_lock_destroy(x) UNUSED(x)
-#ifdef __x86_64__
-static inline void ggml_lock_lock(void* x);
-inline void ggml_lock_lock(void* x) {
-    __asm__ __volatile__("pause\n");
-}
+#if defined(__x86_64__) || defined(_MSC_VER)
+#define ggml_lock_lock(x)    _mm_pause()
 #else
 #define ggml_lock_lock(x)    UNUSED(x)
 #endif