Tested with a 13B model.
You might also want to think about other architectures.

![pause1](https://user-images.githubusercontent.com/3876865/236184685-640fdf7b-3b08-4604-aafd-933ff224a68e.png)
![pause2](https://user-images.githubusercontent.com/3876865/236184698-804275ca-7e8b-40ce-b3d6-14039cfacc3a.png)