@Safari77 2023-05-04
> How does the `eval time (ms / run)` change with this?

Things can be different if CPU would need throttling because 85 °C was reached, but in my case the difference was 60 vs 70 °C.

with pause:
```
llama_print_timings:        load time =   491.01 ms
llama_print_timings:      sample time =   112.94 ms /   381 runs   (    0.30 ms per run)
llama_print_timings: prompt eval time =  2892.00 ms /    27 tokens (  107.11 ms per token)
llama_print_timings:        eval time = 99059.15 ms /   381 runs   (  260.00 ms per run)
llama_print_timings:       total time = 110540.60 ms

Performance counter stats for './main -n 2048 -c 2048 -t 12 --ignore-eos --interactive-first --color --instruct -s 1 -m
 ../gpt4-x-alpaca-13b-native-ggml-q4_0.bin':

      1 220 185,57 msec task-clock:u                     #   11,025 CPUs utilized
                 0      context-switches:u               #    0,000 /sec
                 0      cpu-migrations:u                 #    0,000 /sec
           345 203      page-faults:u                    #  282,910 /sec
 3 145 904 218 579      cpu_core/cycles/u                #    2,578 G/sec                    (98,56%)
 2 769 978 718 128      cpu_atom/cycles/u                #    2,270 G/sec                    (85,03%)
 1 937 590 809 996      cpu_core/instructions/u          #    1,588 G/sec                    (98,56%)
 2 150 075 610 108      cpu_atom/instructions/u          #    1,762 G/sec                    (85,03%)
   114 483 109 780      cpu_core/branches/u              #   93,824 M/sec                    (98,56%)
   113 048 578 832      cpu_atom/branches/u              #   92,649 M/sec                    (85,03%)
       472 475 920      cpu_core/branch-misses/u         #  387,216 K/sec                    (98,56%)
       538 736 029      cpu_atom/branch-misses/u         #  441,520 K/sec                    (85,03%)
18 682 482 544 249      cpu_core/slots:u/                #   15,311 G/sec                    (98,56%)
 2 044 175 523 462      cpu_core/topdown-retiring/u      #     10,9% Retiring                (98,56%)
    35 809 640 118      cpu_core/topdown-bad-spec/u      #      0,2% Bad Speculation         (98,56%)
    80 932 620 092      cpu_core/topdown-fe-bound/u      #      0,4% Frontend Bound          (98,56%)
16 535 620 003 457      cpu_core/topdown-be-bound/u      #     88,4% Backend Bound           (98,56%)
   224 977 900 355      cpu_core/topdown-heavy-ops/u     #      1,2% Heavy Operations       #      9,7% Light Operations        (98,56%)
    20 098 273 461      cpu_core/topdown-br-mispredict/u #      0,1% Branch Mispredict      #      0,1% Machine Clears          (98,56%)
    15 221 484 837      cpu_core/topdown-fetch-lat/u     #      0,1% Fetch Latency          #      0,4% Fetch Bandwidth         (98,56%)
 3 903 133 516 034      cpu_core/topdown-mem-bound/u     #     20,9% Memory Bound           #     67,6% Core Bound              (98,56%)
```

without pause:
```
llama_print_timings:        load time =   494.46 ms
llama_print_timings:      sample time =   113.41 ms /   383 runs   (    0.30 ms per run)
llama_print_timings: prompt eval time =  2892.17 ms /    27 tokens (  107.12 ms per token)
llama_print_timings:        eval time = 100403.52 ms /   383 runs   (  262.15 ms per run)
llama_print_timings:       total time = 105522.82 ms

 Performance counter stats for './main -n 2048 -c 2048 -t 12 --ignore-eos --interactive-first --color --instruct -s 1 -m
 ../gpt4-x-alpaca-13b-native-ggml-q4_0.bin':

      1 236 239,64 msec task-clock:u                     #   11,699 CPUs utilized          
                 0      context-switches:u               #    0,000 /sec                   
                 0      cpu-migrations:u                 #    0,000 /sec                   
           345 993      page-faults:u                    #  279,875 /sec                   
 3 188 758 738 410      cpu_core/cycles/u                #    2,579 G/sec                    (98,53%)
 2 722 421 590 622      cpu_atom/cycles/u                #    2,202 G/sec                    (87,66%)
12 756 258 981 589      cpu_core/instructions/u          #   10,319 G/sec                    (98,53%)
 3 434 808 721 694      cpu_atom/instructions/u          #    2,778 G/sec                    (87,66%)
 2 992 087 196 730      cpu_core/branches/u              #    2,420 G/sec                    (98,53%)
   446 389 930 549      cpu_atom/branches/u              #  361,087 M/sec                    (87,66%)
       474 845 600      cpu_core/branch-misses/u         #  384,105 K/sec                    (98,53%)
       525 726 624      cpu_atom/branch-misses/u         #  425,263 K/sec                    (87,66%)
19 040 516 217 918      cpu_core/slots:u/                #   15,402 G/sec                    (98,53%)
9 876 182 751 945      cpu_core/topdown-retiring/u      #     51,8% Retiring                (98,53%)
    75 459 706 605      cpu_core/topdown-bad-spec/u      #      0,4% Bad Speculation         (98,53%)
   107 405 880 272      cpu_core/topdown-fe-bound/u      #      0,6% Frontend Bound          (98,53%)
 9 011 728 924 394      cpu_core/topdown-be-bound/u      #     47,3% Backend Bound           (98,53%)
    79 615 892 727      cpu_core/topdown-heavy-ops/u     #      0,4% Heavy Operations       #     51,4% Light Operations        (98,53%)
    33 803 030 664      cpu_core/topdown-br-mispredict/u #      0,2% Branch Mispredict      #      0,2% Machine Clears          (98,53%)
    15 939 455 442      cpu_core/topdown-fetch-lat/u     #      0,1% Fetch Latency          #      0,5% Fetch Bandwidth         (98,53%)
 3 402 714 615 342      cpu_core/topdown-mem-bound/u     #     17,8% Memory Bound           #     29,4% Core Bound              (98,53%)
```