@Safari77 2023-05-04
> Interesting, can you suggest an alternative for `arm64` that I can try?

I do not have arm64 to test with, but linux `arch/arm64/include/asm/cmpxchg.h` uses sevl and wfe, try what combination works for you...