@ggerganov 2023-05-04
Tried a few variations of these, but not sure how to measure the effect:

```diff
-- a/ggml.c
+++ b/ggml.c
@@ -11123,7 +11123,21 @@ typedef int ggml_lock_t;
 
 #define ggml_lock_init(x)    UNUSED(x)
 #define ggml_lock_destroy(x) UNUSED(x)
+#ifdef __ARM_NEON
+static inline void ggml_lock_lock(void * x);
+inline void ggml_lock_lock(void * x) {
+    UNUSED(x);
+    for (int i = 0; i < 10; ++i) {
+        __asm __volatile("nop");
+    }
+    __asm __volatile(
+            "  sev\n"
+            "  wfe\n"
+            );
+}
+#else
 #define ggml_lock_lock(x)    UNUSED(x)
+#endif
 #define ggml_lock_unlock(x)  UNUSED(x)
 
 #define GGML_LOCK_INITIALIZER 0
```


Anyway, for x86, if other people can confirm that the performance is not degraded, we can merge the proposed pause.