@Safari77 2023-05-04
> Tried a few variations of these, but not sure how to measure the effect:

Maybe yield is more suitable for arm64. But I used sensors to check out the temperature, and `perf stat -- ./main ...` for performance counter stats.