@slaren 2023-05-05
Testing on a i9 9900k:

PR:
```
llama_print_timings: prompt eval time =  3289.73 ms /   631 tokens (    5.21 ms per token)
llama_print_timings:        eval time = 16655.04 ms /   127 runs   (  131.14 ms per run)
2.10 seconds per pass - ETA 22 minutes
CPU power usage: ~80W
```

Master:
```
llama_print_timings: prompt eval time =  3206.96 ms /   631 tokens (    5.08 ms per token)
llama_print_timings:        eval time = 16859.71 ms /   127 runs   (  132.75 ms per run)
2.17 seconds per pass - ETA 23 minutes
CPU power usage: ~100W
```

For me, this has no downsides, it uses 20% less power and is as fast or faster.